Add files via upload
This project is the **frontend application** of the SGA Localisation system.
It provides a user-friendly interface that allows users to enter their geographic coordinates (latitude and longitude) and instantly obtain the nearest agency or SGA address.  
The interface is built with "Angular", ensuring a responsive, dynamic, and modern user experience.